+++ src/build.rs
@@ -10,13 +10,14 @@ use crate::actions::Action;
 use crate::{CACHE_DIR, ARTIFACTS_DIR, BUILD_CACHE_LIST};
 use crate::entities::auto_version::AutoVersionExtractFromRule;
 use crate::entities::environment::BuildEnvironment;
-use crate::entities::info::ContentInfo;
+use crate::entities::info::{ContentInfo, ShortName};
 use crate::entities::requirements::{Requirement, Satisfy, SatisfyErr};
 use crate::entities::traits::Execute;
 use crate::cmd::{BuildArgs, CleanArgs};
-use crate::configs::DeployerProjectOptions;
+use crate::configs::{DeployerGlobalConfig, DeployerProjectOptions};
 use crate::i18n;
 use crate::pipelines::DescribedPipeline;
+use crate::remote::{sync_to_remote, sync_from_remote, sync_artifacts_from_remote};
 use crate::rw::{copy_all, write, symlink, log, generate_build_log_filepath, build_log};
 use crate::storage::{use_from_storage, add_to_storage};
 use crate::utils::get_current_working_dir;
@@ -154,13 +155,14 @@ fn prepare_build_folder(
 
 pub(crate) fn build(
   config: &mut DeployerProjectOptions,
+  globals: &DeployerGlobalConfig,
   builds: &mut Builds,
   cache_dir: &Path,
   config_dir: &Path,
   storage_dir: &Path,
   args: &BuildArgs,
 ) -> anyhow::Result<()> {
-  if *config == Default::default() { panic!("{}", i18n::CFG_INVALID); }
+  if *config == Default::default() && args.remote_build_folder.is_none() { panic!("{}", i18n::CFG_INVALID); }
   
   if args.link_cache && args.copy_cache { panic!(
     "Select only one option from `{}` and `{}`. See help via `{}`.", "c".green(), "C".green(), "deployer build -h".green()
@@ -177,9 +179,90 @@ pub(crate) fn build(
   if args.silent && args.no_pipe { panic!(
     "Select only one option from `{}` and `{}`. See help via `{}`.", "s".green(), "t".green(), "deployer build -h".green()
   ); }
-
+  if args.remote_build_folder.is_some() && args.pipeline_tags.is_empty() { panic!(
+    "You always should specify Pipeline tags for executing while remote builds."
+  ) }
+  if !args.remote_host_short_names.is_empty() && (
+    args.remote_build_folder.is_some() ||
+    args.pipeline_tags.is_empty() ||
+    args.link_cache ||
+    args.copy_cache ||
+    args.fresh ||
+    args.build_at.is_some() ||
+    args.current
+  ) { panic!(
+    "If you specify remote hosts to build on, you should specify only Pipelines list, `{}`/`{}` flags and nothing more.",
+    "silent".green(),
+    "no_pipe".green(),
+  ) }
+  
+  if let Some(build_dir) = &args.remote_build_folder && !args.pipeline_tags.is_empty() {
+    let config = crate::rw::read::<DeployerProjectOptions>(build_dir, crate::PROJECT_CONF);
+    let artifacts_dir = prepare_artifacts_folder(build_dir)?;
+    
+    for pipeline_tag in &args.pipeline_tags {
+      if let Some(pipeline) = &config.pipelines.iter().find(|p| p.title.as_str().eq(pipeline_tag)) {
+        let env = BuildEnvironment {
+          build_dir,
+          cache_dir,
+          config_dir,
+          storage_dir,
+          artifacts_dir: &artifacts_dir,
+          new_build: true,
+          silent_build: false,
+          no_pipe: true,
+          ignore: &config.cache_files,
+          remotes: &globals.remote_hosts,
+        };
+        
+        execute_pipeline(&config, env, pipeline)?;
+        enplace_artifacts(&config, env, false)?;
+      } else {
+        panic!(
+          "There is no such Pipeline `{}` set up for this project. Maybe, you've forgotten set up this Pipeline for project via `{}`?",
+          pipeline_tag.green(),
+          "deployer with {pipeline-short-name-and-ver}".green(),
+        );
+      }
+    }
+    
+    return Ok(())
+  }
+  
   let curr_dir = std::env::current_dir().expect("Can't get current dir!");
   let artifacts_dir = prepare_artifacts_folder(&curr_dir)?;
+  
+  if !args.remote_host_short_names.is_empty() {
+    let mut remote = vec![];
+    for short_name in &args.remote_host_short_names {
+      if let Some(host) = globals.remote_hosts.get(&ShortName::new(short_name)?) {
+        remote.push(host.clone());
+      }
+    }
+    
+    for pipeline_tag in &args.pipeline_tags {
+      if let Some(pipeline) = config.pipelines.iter().find(|p| p.title.as_str().eq(pipeline_tag)) {
+        let (build_path, _) = prepare_build_folder(config, builds, pipeline, &curr_dir, cache_dir, args)?;
+        
+        for host in remote.iter() {
+          println!("{} `{}`...", i18n::START_BUILD_AT_REMOTE, host.short_name.as_str().green());
+          let now = std::time::Instant::now();
+          let generated_remote = sync_to_remote(&build_path, host, &config.cache_files)?;
+          if let Err(e) = host.call_deployer_to_build(&generated_remote, pipeline.title.as_str()) { println!("{}", e); };
+          sync_artifacts_from_remote(&generated_remote, &artifacts_dir, host)?;
+          println!("{} `{}` ({}).", i18n::BUILT_AT_REMOTE, host.short_name.as_str().green(), format!("{:.2?}", now.elapsed()).green());
+        }
+      } else {
+        panic!(
+          "There is no such Pipeline `{}` set up for this project. Maybe, you've forgotten set up this Pipeline for project via `{}`?",
+          pipeline_tag.green(),
+          "deployer with {pipeline-short-name-and-ver}".green(),
+        );
+      }
+    }
+    
+    return Ok(())
+  }
 
   if args.pipeline_tags.is_empty() {
     if config.pipelines.is_empty() {
@@ -205,11 +288,12 @@ pub(crate) fn build(
         new_build,
         silent_build: args.silent,
         no_pipe: args.no_pipe,
+        ignore: &config.cache_files,
+        remotes: &globals.remote_hosts,
       };
       
       execute_pipeline(config, env, pipeline)?;
-      
-      enplace_artifacts(config, env, true)?;
+      enplace_artifacts(config, env, false)?;
     }
   } else {
     for pipeline_tag in &args.pipeline_tags {
@@ -229,11 +313,12 @@ pub(crate) fn build(
           new_build,
           silent_build: args.silent,
           no_pipe: args.no_pipe,
+          ignore: &config.cache_files,
+          remotes: &globals.remote_hosts,
         };
         
         execute_pipeline(config, env, pipeline)?;
-        
-        enplace_artifacts(config, env, true)?;
+        enplace_artifacts(config, env, false)?;
       } else {
         panic!(
           "There is no such Pipeline `{}` set up for this project. Maybe, you've forgotten set up this Pipeline for project via `{}`?",
@@ -253,7 +338,24 @@ pub(crate) fn execute_pipeline(
   pipeline: &DescribedPipeline,
 ) -> anyhow::Result<()> {
   use std::io::{stdout, Write};
-  use std::time::Instant;
+  use std::time::{Instant, Duration};
+  
+  let mut total_time = Duration::from_secs(0);
+  let log_file = generate_build_log_filepath(
+    &config.project_name,
+    &pipeline.title,
+    env.cache_dir,
+  );
+  
+  if !env.silent_build { println!("{}", i18n::STARTING_PIPELINE.replace("{}", &pipeline.title)); }
+  build_log(&log_file, &[format!("Starting the `{}` Pipeline...", pipeline.title)])?;
+  
+  let canonicalized = env.build_dir.canonicalize()?;
+  let canonicalized = canonicalized.to_str().expect("Can't convert `Path` to string!");
+  if !env.silent_build { println!("{}: {}", i18n::BUILD_PATH, canonicalized); }
+  build_log(&log_file, &[format!("{}: {}", i18n::BUILD_PATH, canonicalized)])?;
+  
+  let now = Instant::now();
   
   #[allow(clippy::mutable_key_type)]
   let mut requirements = HashSet::<Requirement>::new();
@@ -278,19 +380,12 @@ pub(crate) fn execute_pipeline(
     return Ok(())
   }
   
-  let log_file = generate_build_log_filepath(
-    &config.project_name,
-    &pipeline.title,
-    env.cache_dir,
-  );
-  
-  if !env.silent_build { println!("{}", i18n::STARTING_PIPELINE.replace("{}", &pipeline.title)); }
-  build_log(&log_file, &[format!("Starting the `{}` Pipeline...", pipeline.title)])?;
-  
-  let canonicalized = env.build_dir.canonicalize()?;
-  let canonicalized = canonicalized.to_str().expect("Can't convert `Path` to string!");
-  if !env.silent_build { println!("{}: {}", i18n::BUILD_PATH, canonicalized); }
-  build_log(&log_file, &[format!("{}: {}", i18n::BUILD_PATH, canonicalized)])?;
+  if !requirements.is_empty() {
+    let elapsed = now.elapsed();
+    println!("{} {}.", i18n::REQ_CHECKS_TOOK, format!("{:.2?}", elapsed).green());
+    build_log(&log_file, &[format!("{} {:.2?}.", i18n::REQ_CHECKS_TOOK, elapsed)])?;
+    total_time += elapsed;
+  }
   
   let mut cntr = 1usize;
   let total = pipeline.actions.len();
@@ -307,6 +402,20 @@ pub(crate) fn execute_pipeline(
     let now = Instant::now();
     
     let (status, output) = match &action.action {
+      Action::SyncToRemote(remote_name) => {
+        if let Some(remote) = env.remotes.get(remote_name) {
+          if let Err(e) = sync_to_remote(env.build_dir, remote, env.ignore) { (false, vec![e.to_string()]) } else { (true, vec![]) }
+        } else {
+          (false, vec![i18n::NO_SUCH_REMOTE.to_string()])
+        }
+      },
+      Action::SyncFromRemote(remote_name) => {
+        if let Some(remote) = env.remotes.get(remote_name) {
+          if let Err(e) = sync_from_remote(env.build_dir, remote) { (false, vec![e.to_string()]) } else { (true, vec![]) }
+        } else {
+          (false, vec![i18n::NO_SUCH_REMOTE.to_string()])
+        }
+      },
       Action::Custom(cmd) => cmd.execute(env)?,
       Action::Check(check) => check.execute(env)?,
       Action::PreBuild(a) | Action::Build(a) | Action::PostBuild(a) | Action::Test(a) => a.execute(env)?,
@@ -364,6 +473,7 @@ pub(crate) fn execute_pipeline(
     };
     
     let elapsed = now.elapsed();
+    total_time += elapsed;
     if !env.no_pipe { build_log(&log_file, &output)?; }
     build_log(&log_file, &[
       format!(
@@ -379,7 +489,7 @@ pub(crate) fn execute_pipeline(
     if !env.silent_build {
       if !env.no_pipe {
         println!("{} ({}).", status_str, format!("{:.2?}", elapsed).green());
-        for line in output { println!("{}", line); }
+        for line in &output { println!("{}", line); }
       } else {
         println!("[{}/{}] {} -{} ({}).", cntr, total, i18n::STARTING_ACTION.replace("{}", &action.title.blue().italic()), status_str, format!("{:.2?}", elapsed).green());
       }
@@ -389,6 +499,9 @@ pub(crate) fn execute_pipeline(
     
     if !status { return Ok(()) }
   }
+  
+  println!("{} {}.", i18n::DONE_IN, format!("{:.2?}", total_time).green());
+  build_log(&log_file, &[format!("{} {:.2?}.", i18n::DONE_IN, total_time)])?;
   
   Ok(())
 }