+++ src/entities/remote_host.rs
@@ -4,7 +4,6 @@ use std::borrow::Cow;
 use std::net::IpAddr;
 use std::path::{Path, PathBuf};
 use std::sync::Arc;
-use std::time::Duration;
 use tokio::net::ToSocketAddrs;
 
 use crate::entities::info::ShortName;
@@ -20,13 +19,45 @@ pub(crate) struct RemoteHost {
 
 impl RemoteHost {
   pub(crate) fn check(&self) -> anyhow::Result<()> {
+    const PKG_NAME: &str = env!("CARGO_PKG_NAME");
+    const PKG_VERSION: &str = env!("CARGO_PKG_VERSION");
+    
+    let shell = match std::env::var("DEPLOYER_SH_PATH") {
+      Ok(path) => path,
+      Err(_) => "/bin/bash".to_string(),
+    };
+    
     let rt = tokio::runtime::Runtime::new()?;
     rt.block_on(async {
       let mut session = Session::connect(&self.ssh_private_key_file, &self.username, (self.ip, self.port)).await?;
-      let (status, out) = session.call("which deployer").await?;
+      let (s, o) = session.call(&format!(r#"{} -c "~/.cargo/bin/deployer -V""#, shell)).await?;
       session.close().await?;
-      if status == 0 && !out.is_empty() && out.contains("deployer") { Ok(()) }
-      else { bail!("Remote host doesn't contains `deployer` executable in PATH.") }
+      if s != 0 || o.is_empty() || !o.contains(&format!("{} {}", PKG_NAME, PKG_VERSION)) {
+        bail!(r#"Deployer version on remote host didn't match with this Deployer version (out: "{}")"#, o.trim()) }
+      else { Ok(()) }
+    })
+  }
+  
+  pub(crate) fn call_deployer_to_build(&self, remote_build_dir: &Path, pipeline: &str) -> anyhow::Result<()> {
+    let shell = match std::env::var("DEPLOYER_SH_PATH") {
+      Ok(path) => path,
+      Err(_) => "/bin/bash".to_string(),
+    };
+    
+    let rt = tokio::runtime::Runtime::new()?;
+    rt.block_on(async {
+      let mut session = Session::connect(&self.ssh_private_key_file, &self.username, (self.ip, self.port)).await?;
+      let (s, o) = session.call(
+        &format!(
+          r#"{} -c "~/.cargo/bin/deployer build -r {} {}""#,
+          shell,
+          remote_build_dir.to_string_lossy().as_str(),
+          pipeline,
+        )
+      ).await?;
+      session.close().await?;
+      if s != 0 { bail!("{}", o) }
+      else { Ok(()) }
     })
   }
   
@@ -50,7 +81,15 @@ impl RemoteHost {
 }
 
 struct Client {}
-impl russh::client::Handler for Client { type Error = anyhow::Error; }
+
+#[async_trait::async_trait]
+impl russh::client::Handler for Client {
+  type Error = anyhow::Error;
+  
+  async fn check_server_key(&mut self, _: &russh::keys::ssh_key::PublicKey) -> Result<bool, Self::Error> {
+    Ok(true)
+  }
+}
 
 pub(crate) struct Session {
   session: russh::client::Handle<Client>,
@@ -62,7 +101,6 @@ impl Session {
     
     let key_pair = load_secret_key(key_path, None)?;
     let config = russh::client::Config {
-      inactivity_timeout: Some(Duration::from_secs(5)),
       preferred: russh::Preferred {
         kex: Cow::Owned(vec![russh::kex::DH_G14_SHA256]),
         ..Default::default()
@@ -73,8 +111,11 @@ impl Session {
     let config = Arc::new(config);
     let sh = Client {};
 
-    let mut session = russh::client::connect(config, addrs, sh).await?;
-    let auth_res = session.authenticate_publickey(user, PrivateKeyWithHashAlg::new(Arc::new(key_pair), None)?).await?;
+    let mut session = match russh::client::connect(config, addrs, sh).await {
+      Ok(s) => s,
+      Err(e) => bail!("Client connect failed: {e:?}"),
+    };
+    let auth_res = session.authenticate_publickey(user, PrivateKeyWithHashAlg::new(Arc::new(key_pair), Some(russh::keys::HashAlg::Sha256))?).await?;
 
     if !auth_res { bail!("Authentication failed: {auth_res:?}"); }
 
@@ -100,7 +141,7 @@ impl Session {
     }
     
     let out = String::from_utf8_lossy_owned(out_buf);
-    Ok((status.expect("Remote program did not exit cleanly."), out))
+    Ok((status.unwrap_or(0), out))
   }
 
   async fn close(&mut self) -> anyhow::Result<()> {