+++ src/main.rs
@@ -123,6 +123,7 @@ fn main() {
   
   // Чтение конфигов
   let mut globals = read::<DeployerGlobalConfig>(&config_folder, GLOBAL_CONF);
+  DeployerGlobalConfig::make_sure_contain_defaults(&mut globals.actions_registry);
   let mut config = read::<DeployerProjectOptions>(&get_current_working_dir().unwrap(), PROJECT_CONF);
   if config == Default::default() { config = read::<DeployerProjectOptions>(&get_current_working_dir().unwrap(), HIDDEN_PROJECT_CONF); }
   let mut builds = read::<Builds>(&cache_folder, BUILD_CACHE_LIST);
@@ -193,7 +194,7 @@ fn main() {
       write(get_current_working_dir().unwrap(), PROJECT_CONF, &config);
     },
     DeployerExecType::Build(args) => {
-      build(&mut config, &mut builds, &cache_folder, &config_folder, &storage_folder, &args).unwrap();
+      build(&mut config, &globals, &mut builds, &cache_folder, &config_folder, &storage_folder, &args).unwrap();
       write(&cache_folder, BUILD_CACHE_LIST, &builds);
     },
     DeployerExecType::Clean(args) => {