+++ CHANGELOG.md
@@ -7,6 +7,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.0] - 2025-01-14
+
+### Added
+
+- `SyncToRemote` and `SyncFromRemote` Actions to sync build folders between current and remote hosts.
+- Remote Action and Pipeline execution support.
+- `-r` and `-R` arguments to `build` command.
+
+### Changed
+
+- Documentation and `i18n`.
+
 ## [1.1.0-beta-2] - 2025-01-14
 
 ### Fixed
@@ -267,6 +279,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Build support.
 - TUI.
 
+[1.2.0]:          https://github.com/impulse-sw/deployer/compare/1.1.0-beta-1...1.2.0
 [1.1.0-beta-2]:   https://github.com/impulse-sw/deployer/compare/1.1.0-beta-1...1.1.0-beta-2
 [1.1.0-beta-1]:   https://github.com/impulse-sw/deployer/compare/1.0.1...1.1.0-beta-1
 [1.0.1]:          https://github.com/impulse-sw/deployer/compare/1.0.0...1.0.1